requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
pydantic>=2.0.0
prometheus-client>=0.19.0
